Clean up RouteGenerator
* Split classes up into their own files for readability and to permit
  class autoloading.

* Move route-related classes into "Routes" namespace for organizational
  purposes.

* Remove improper use of references. See #317.

* Add typehints to class method interfaces and class attributes.
